Move the entity helpers and the texture decoder hook out of Utils
