+++ GUI/Forms/AboutForm.cs
@@ -7,6 +7,7 @@ using GUI.Types.GLViewers;
 using GUI.Utils;
 using Svg.Skia;
 using ValveResourceFormat.Renderer;
+using ValveResourceFormat.TextureDecoders;
 
 namespace GUI.Forms
 {
