+++ Renderer/Renderer/AnimationController.IK.cs
@@ -1,3 +1,4 @@
+using ValveResourceFormat.ResourceTypes;
 using ValveResourceFormat.ResourceTypes.ModelAnimation;
 
 namespace ValveResourceFormat.Renderer
