+++ Renderer/Renderer/Camera.cs
@@ -1,3 +1,5 @@
+using ValveResourceFormat.ResourceTypes;
+
 namespace ValveResourceFormat.Renderer
 {
     /// <summary>
