+++ ValveResourceFormat/Particles/Initializers/RemapInitialTransformDirectionToRotation.cs
@@ -1,3 +1,5 @@
+using ValveResourceFormat.ResourceTypes;
+
 namespace ValveResourceFormat.Particles.Initializers
 {
     /// <summary>
