+++ ValveResourceFormat/Particles/Particle.cs
@@ -1,3 +1,4 @@
+using ValveResourceFormat.ResourceTypes;
 using ValveResourceFormat.Serialization.KeyValues;
 
 namespace ValveResourceFormat.Particles
